Add function to clear completed to do list
