+++ src/index.html
@@ -23,7 +23,7 @@
 
       <ul id="list"></ul>
 
-      <button class="clear">Clear all completed</button>
+      <button class="clear" type="button">Clear all completed</button>
     </section>
 
     <script src="./index.js" type="module"></script>
