+++ src/modules/localStorage.js
@@ -1,0 +1,61 @@
+let todolists = [];
+
+class LocalStorage {
+  static getToDoLists() {
+    if (localStorage.getItem('todolists') === null) {
+      localStorage.setItem('todolists', JSON.stringify(todolists));
+    } else {
+      todolists = JSON.parse(localStorage.getItem('todolists'));
+    }
+    return todolists;
+  }
+
+  static addToDoLists(item) {
+    const todolists = LocalStorage.getToDoLists();
+    todolists.push(item);
+    localStorage.setItem('todolists', JSON.stringify(todolists));
+  }
+
+  static removeToDoLists(newdesc) {
+    const todolists = LocalStorage.getToDoLists();
+    todolists.forEach((item, index) => {
+      if (item.desc === newdesc) {
+        todolists.splice(index, 1);
+        localStorage.setItem('todolists', JSON.stringify(todolists));
+      }
+    });
+
+    // RESET INDEX
+    LocalStorage.resetIndex();
+  }
+
+  static resetIndex() {
+    const todolists = LocalStorage.getToDoLists();
+    todolists.forEach((item, index) => {
+      item.index = index + 1;
+      localStorage.setItem('todolists', JSON.stringify(todolists));
+    });
+  }
+
+  static updateDesc(desc, index) {
+    const todolists = LocalStorage.getToDoLists();
+    for (let i = 0; i < todolists.length; i += 1) {
+      if (todolists[i].index === +index) {
+        todolists[i].desc = desc;
+        localStorage.setItem('todolists', JSON.stringify(todolists));
+      }
+    }
+  }
+
+  static taskComp(index) {
+    const todolists = LocalStorage.getToDoLists();
+    for (let i = 0; i < todolists.length; i += 1) {
+      if (todolists[i].index === +index) {
+        todolists[i].comp = !todolists[i].comp;
+      }
+    }
+    localStorage.setItem('todolists', JSON.stringify(todolists));
+  }
+}
+
+export default LocalStorage;
